requirements.txt
# Core dependencies for RSS Telegram Bot
feedparser==6.0.11
boto3==1.35.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0

# Testing dependencies
pytest==8.3.3
hypothesis==6.112.1
moto==5.0.0
PyYAML==6.0.2

# Development dependencies
ruff==0.6.9